# src/main/java/org/example/CustomersResource.java
package org.example;

import jakarta.ws.rs.*;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;

import java.util.HashMap;
import java.util.Map;
import java.util.concurrent.atomic.AtomicInteger;

@Path("/customers")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
public class CustomersResource {

    private final Map<Integer, Customer> customers = new HashMap<>();
    private final AtomicInteger idCounter = new AtomicInteger(1);

    @GET
    public Response getCustomers() {
        return Response.ok(customers).build();
    }

    @POST
    public Response addCustomer(Customer customer) {
        Integer newId = idCounter.getAndIncrement();
        customer.setId(newId);
        customers.put(newId, customer);
        return Response.status(Response.Status.CREATED).entity(customer).build();
    }

    @DELETE
    @Path("/{id}")
    public Response deleteCustomer(@PathParam("id") Integer id) {
        if (customers.containsKey(id)) {
            customers.remove(id);
            return Response.noContent().build();
        }
        return Response.status(Response.Status.NOT_FOUND).build();
    }
}